refactor: move learner home to separate app
