--- lms/djangoapps/learner_dashboard/README.rst
@@ -1,21 +1,0 @@
-=================
-Learner dashboard
-=================
-
-This djangoapp houses 2 related dashboards owned and developed, currently by 2 separate teams:
-
-1. Courses Dashboard
-2. Programs Dashboard
-
-Courses Dashboard
-=================
-
-The Courses section of the Learner Dashboard is a backend supporting a new MFE experience of the learner dashboard.
-
-This aims to replace the existing dashboard at::
-    /common/djangoapps/student/views/dashboard.py
-
-Programs Dashboard
-==================
-
-See :ref:`programs`.rst doc.
