+++ lms/djangoapps/learner_dashboard/urls.py
@@ -2,15 +2,9 @@
 
 from django.urls import path, re_path
 
-from lms.djangoapps.learner_dashboard import learner_views, programs, program_views
+from lms.djangoapps.learner_dashboard import programs, program_views
 
-# Learner Dashboard Routing
 urlpatterns = [
-    path('learner/', learner_views.dashboard_view, name='dashboard_view')
-]
-
-# Program Dashboard Routing
-urlpatterns += [
     path('programs/', program_views.program_listing, name='program_listing_view'),
     re_path(r'^programs/(?P<program_uuid>[0-9a-f-]+)/$', program_views.program_details, name='program_details_view'),
     re_path(r'^programs/(?P<program_uuid>[0-9a-f-]+)/discussion/$', program_views.ProgramDiscussionIframeView.as_view(),
