+++ lms/djangoapps/learner_home/README.md
@@ -1,0 +1,8 @@
+=================
+Learner Home
+=================
+
+This is the new dashboard for learner courses, built as a backend supporting a new MFE experience of the "student dashboard".
+
+This aims to replace the existing dashboard at::
+    /common/djangoapps/student/views/dashboard.py
