+++ lms/djangoapps/learner_home/serializers.py
