+++ lms/urls.py
@@ -195,6 +195,9 @@ urlpatterns = [
     path('dashboard/', include('lms.djangoapps.learner_dashboard.urls')),
     path('api/dashboard/', include('lms.djangoapps.learner_dashboard.api.urls', namespace='dashboard_api')),
 
+    # Learner Home
+    path('learner/', include ('lms.djangoapps.learner_home.urls', namespace='learner_home')),
+
     path(
         'api/experiments/',
         include(
